Added admin LogIn/logOut functionality
