+++ .settings/org.eclipse.wst.common.project.facet.core.xml
@@ -3,6 +3,6 @@
   <runtime name="Apache Tomcat v9.0"/>
   <fixed facet="wst.jsdt.web"/>
   <installed facet="wst.jsdt.web" version="1.0"/>
-  <installed facet="java" version="1.8"/>
   <installed facet="jst.web" version="3.1"/>
+  <installed facet="java" version="1.8"/>
 </faceted-project>
