+++ src/main/webapp/WEB-INF/resources/css/carousel.css
@@ -46,8 +46,8 @@ body {
 /* Since positioning the image, we need to help out the caption */
 .carousel-caption {
   z-index: 10;
-
-  
+  top: 230px;
+  bottom: auto;
 }
 
 /* Declare heights because of positioning of img element */
