+++ src/main/webapp/WEB-INF/resources/css/main.css
@@ -33,3 +33,13 @@
 	white-space: nowrap;
 }
 
+.login-box {
+    text-align: center;
+}
+
+.login-container {
+    width: 40%;
+    margin: auto;
+    margin-top: 100px;
+}
+
